displaying the articles works
