+++ app/src/pages/HomePage.js
@@ -1,28 +1,13 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import NavBar from '../components/NavBar';
-import { getAllArticles } from '../utils/functions/articles';
-
-const INITIAL_STATE = {
-    articles: {}
-};
+import ArticleList from '../components/Homepage/ArticleList';
 
 export default function HomePage() {
-    const [state, setState] = useState({
-        ...INITIAL_STATE
-    });
-
-    useEffect(() => {
-        getAllArticles().then((articles) => {
-            setState({ articles: articles.article_list });
-        }).catch((err) => {
-            console.log(err);
-        })
-    });
-
     return (
         <div id="home-page">
             <NavBar />
-            <p>this is the new home page</p>
+            <p>This is the new home page</p>
+            <ArticleList />
         </div>
     );
 }
